fix: extract most of `[stackId]/+layout.svelte` contents into own component
